# Ops: recovery verification report - hotfix functionality confirmed