--- a/artifacts/verify.md
+++ b/artifacts/verify.md
@@ -1,7 +1,7 @@
 # Verification Report
 
-Generated: 2025-08-25T08:24:43.115Z
-Duration: 18.7s
+Generated: 2025-08-25T10:49:50.307Z
+Duration: 18.1s
 Status: ❌ Some verification steps failed
 
 **Failed Steps**: 2