--- a/src/providers/llm-anthropic.ts
+++ b/src/providers/llm-anthropic.ts
@@ -3,7 +3,7 @@ import type { LLM } from './index.js';
 const AnthropicProvider: LLM = {
   name: 'anthropic',
   async complete({ prompt, system, temperature }) {
-    const mod: any = await import('@anthropic-ai/sdk');
+    const mod: any = await eval('import("@anthropic-ai/sdk")');
     const client = new mod.default({ apiKey: process.env.ANTHROPIC_API_KEY });
     const res = await client.messages.create({
       model: process.env.ANTHROPIC_MODEL ?? 'claude-3-5-sonnet-20240620',