--- a/src/providers/llm-openai.ts
+++ b/src/providers/llm-openai.ts
@@ -1,32 +1,20 @@
-import type { LLM } from './index';
+import type { LLM } from './index.js';
 
-const OpenAI: LLM = {
+const OpenAIProvider: LLM = {
   name: 'openai',
-  async complete({ prompt, system, temperature = 0.7 }) {
-    try {
-      const { OpenAI: OpenAISDK } = await import('openai');
-      const client = new OpenAISDK({
-        apiKey: process.env.OPENAI_API_KEY
-      });
-
-      const messages: Array<{ role: 'system' | 'user'; content: string }> = [];
-      if (system) {
-        messages.push({ role: 'system', content: system });
-      }
-      messages.push({ role: 'user', content: prompt });
-
-      const response = await client.chat.completions.create({
-        model: 'gpt-3.5-turbo',
-        messages,
-        temperature,
-        max_tokens: 1000
-      });
-
-      return response.choices[0]?.message?.content || '[no response]';
-    } catch (error) {
-      throw new Error(`OpenAI API error: ${error instanceof Error ? error.message : 'Unknown error'}`);
-    }
+  async complete({ prompt, system, temperature }) {
+    const mod: any = await import('openai');
+    const client = new mod.default({ apiKey: process.env.OPENAI_API_KEY });
+    const res = await client.chat.completions.create({
+      model: process.env.OPENAI_MODEL ?? 'gpt-4o-mini',
+      messages: [
+        ...(system ? [{ role: 'system', content: system }] : []),
+        { role: 'user', content: prompt }
+      ],
+      temperature: temperature ?? 0
+    });
+    return res?.choices?.[0]?.message?.content ?? '';
   }
 };
 
-export default OpenAI;
+export default OpenAIProvider;