--- a/src/providers/llm-openai.ts
+++ b/src/providers/llm-openai.ts
@@ -3,7 +3,7 @@ import type { LLM } from './index.js';
 const OpenAIProvider: LLM = {
   name: 'openai',
   async complete({ prompt, system, temperature }) {
-    const mod: any = await import('openai');
+    const mod: any = await eval('import("openai")');
     const client = new mod.default({ apiKey: process.env.OPENAI_API_KEY });
     const res = await client.chat.completions.create({
       model: process.env.OPENAI_MODEL ?? 'gpt-4o-mini',